all fix need to do coupon-service

+++ category-service/src/main/java/com/maorzehavi/categoryservice/service/CategoryService.java
@@ -1,11 +1,10 @@
 package com.maorzehavi.categoryservice.service;
 
-import com.maorzehavi.categoryservice.model.dto.request.CategoryRequest;
-import com.maorzehavi.categoryservice.model.dto.request.CategoryResponse;
 import com.maorzehavi.categoryservice.model.entity.Category;
 import com.maorzehavi.categoryservice.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,34 +14,37 @@ import java.util.Optional;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final RestTemplate restTemplate;
 
-    public Optional<CategoryResponse> findById(Long id) {
-        return categoryRepository.findById(id)
-                .map(this::mapToResponse);
-    }
-
-    public CategoryResponse test(CategoryRequest categoryRequest) {
-        return mapToResponse(categoryRepository.save(mapToEntity(categoryRequest)));
-    }
-
-    public Optional<CategoryResponse> createCategory(CategoryRequest categoryRequest) {
-        if (categoryRepository.existsByName(categoryRequest.getName())) {
-            throw new IllegalArgumentException("Category with name " + categoryRequest.getName() + " already exists");
+    public Optional<Category> create(String name) {
+        if (existsByName(name)) {
+            return Optional.empty();
         }
-        Category category = mapToEntity(categoryRequest);
-        Category savedCategory = categoryRepository.save(category);
-        return Optional.of(mapToResponse(savedCategory));
+        Category category = Category.builder()
+                .name(name)
+                .build();
+        return Optional.of(categoryRepository.save(category));
     }
 
-    public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
+    public Optional<Category> getCategory(Long id) {
+        return categoryRepository.findById(id);
     }
 
-    public List<CategoryResponse> findAll() {
-        return categoryRepository.findAll()
-                .stream()
-                .map(this::mapToResponse)
-                .toList();
+    public List<Category> getAllCategories() {
+        return categoryRepository.findAll();
+    }
+
+    public boolean deleteCategory(Long id) {
+        boolean exists = Boolean.TRUE.equals(restTemplate.getForObject("http://coupon-service/api/v1/coupons/category/exists/" + id, Boolean.class));
+        if (exists) {
+            throw new RuntimeException("Cannot delete category with coupons attached");
+        }
+
+        if (existsById(id)) {
+            categoryRepository.deleteById(id);
+            return true;
+        }
+        throw new RuntimeException("Category with id " + id + " not found");
     }
 
     public boolean existsById(Long id) {
@@ -50,16 +52,9 @@ public class CategoryService {
     }
 
 
-    private Category mapToEntity(CategoryRequest categoryRequest) {
-        return Category.builder()
-                .name(categoryRequest.getName())
-                .build();
+    private boolean existsByName(String name) {
+        return categoryRepository.existsByName(name);
     }
 
-    private CategoryResponse mapToResponse(Category category) {
-        return CategoryResponse.builder()
-                .id(category.getId())
-                .name(category.getName())
-                .build();
-    }
+
 }
